Fixed write-policy-dir. Made spome fields in the template optional. Fixed up new unit tests in the sane multifolder structure. Removed crud flag.
