+++ examples/input-dir/policy-puma.yml
@@ -1,5 +1,5 @@
 policy_with_crud_levels:
-- name: 'RoleNameWithCRUD'
+  name: 'RoleNameWithCRUD'
   description: 'Why I need these privs'
   role_arn: 'arn:aws:iam::123456789012:role/RiskyEC2'
   read:
