+++ setup.cfg
@@ -1,0 +1,4 @@
+[nosetests]
+exe = True
+tests = test/, test/analysis/, test/command/, test/querying/, test/util/, test/writing/
+verbosity=2
